create composion.hpp
